+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Extend the "Settings (eHoldings): View settings" permission to view the "Root proxy" settings and "Knowledge Base" details. (UIEH-1355)
 * Don't display the deleted custom label. (UIEH-1364)
 * Revise the 'Settings (eHoldings): View settings' permission. (UIEH-1355)
+* Display custom labels while editing. (UIEH-1364)
 
 ## [8.0.1] (https://github.com/folio-org/ui-eholdings/tree/v8.0.1) (2023-03-13)